[COZY-77] refactor: 피드백 반영 및 파일 구조 변경
