+++ src/main/java/com/cozymate/cozymate_server/domain/mate/Mate.java
@@ -5,6 +5,8 @@ import com.cozymate.cozymate_server.domain.member.Member;
 import com.cozymate.cozymate_server.domain.room.Room;
 import com.cozymate.cozymate_server.global.utils.BaseTimeEntity;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -36,6 +38,7 @@ public class Mate extends BaseTimeEntity {
     private boolean isExit = false;
 
     // 룸메이트 대기상태 조회를 위해 EntryStatus를 추가하였습니다.
+    @Enumerated(EnumType.STRING)
     private EntryStatus entryStatus = EntryStatus.PENDING;
 
     private boolean isRoomManager = false;
