+++ src/main/java/com/cozymate/cozymate_server/domain/mate/repository/MateRepository.java
@@ -1,5 +1,6 @@
-package com.cozymate.cozymate_server.domain.mate;
+package com.cozymate.cozymate_server.domain.mate.repository;
 
+import com.cozymate.cozymate_server.domain.mate.Mate;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MateRepository extends JpaRepository<Mate, Long> {
