+++ src/main/java/com/cozymate/cozymate_server/domain/room/Room.java
@@ -5,6 +5,8 @@ import com.cozymate.cozymate_server.domain.room.enums.RoomStatus;
 import com.cozymate.cozymate_server.global.utils.BaseTimeEntity;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -38,16 +40,12 @@ public class Room extends BaseTimeEntity {
     // s3에서 /persona/1 과 같은 식으로 가져오는게 좋아보입니다.
     private Integer profileImage;
 
+    @Enumerated(EnumType.STRING)
     private RoomStatus status = RoomStatus.WAITING;
 
     // 화면상에서 베스트 룸메이트를 연관 시킬 필요가 없어보여서 이름만 저장하도록 했습니다.
     private String bestMateName;
 
     private int numOfArrival = 1;
-
-    // room <-> mate OneToMany 매핑
-    @OneToMany(mappedBy = "room", fetch = FetchType.LAZY, cascade = CascadeType.ALL)
-    private List<Mate> mateList;
-
 
 }
