+++ src/main/java/com/cozymate/cozymate_server/domain/room/controller/RoomController.java
@@ -1,5 +1,7 @@
-package com.cozymate.cozymate_server.domain.room;
+package com.cozymate.cozymate_server.domain.room.controller;
 
+import com.cozymate.cozymate_server.domain.room.Room;
+import com.cozymate.cozymate_server.domain.room.service.RoomCommandService;
 import com.cozymate.cozymate_server.domain.room.dto.RoomCreateRequest;
 import com.cozymate.cozymate_server.domain.room.dto.RoomCreateResponse;
 import com.cozymate.cozymate_server.global.response.ApiResponse;
@@ -18,16 +20,13 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/rooms")
 public class RoomController {
 
-    private final RoomService roomService;
+    private final RoomCommandService roomCommandService;
 
-    /**
-     * [POST] 방 생성
-     */
     @PostMapping("/create")
     @Operation(summary = "[바니] 방생성 기능", description = "방이름, 프로필이미지, 인원수를 입력합니다.")
     public ResponseEntity<ApiResponse<RoomCreateResponse>> createRoom(@Valid @RequestBody RoomCreateRequest request) {
         // TODO: 시큐리티 이용해 사용자 인증 받아야 함.
-        Room room = roomService.createRoom(request);
+        Room room = roomCommandService.createRoom(request);
         RoomCreateResponse response = new RoomCreateResponse(room.getId(), room.getProfileImage(), room.getInviteCode());
         return ResponseEntity.status(SuccessStatus._OK.getHttpStatus())
             .body(ApiResponse.onSuccess(response));
