+++ src/main/java/com/cozymate/cozymate_server/domain/room/dto/RoomCreateRequest.java
@@ -26,7 +26,7 @@ public class RoomCreateRequest {
     @Min(2)
     @Max(6)
     private Integer maxMateNum;
-//   추후 수정 예정, 테스트 위해 임시로 둔 값
+//   TODO: 추후 수정 예정, 테스트 위해 임시로 둔 값
     @NotNull
     private Long creatorId;
