+++ src/main/java/com/cozymate/cozymate_server/domain/room/service/RoomQueryService.java
@@ -1,0 +1,12 @@
+package com.cozymate.cozymate_server.domain.room.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Transactional(readOnly = true)
+@Service
+@RequiredArgsConstructor
+public class RoomQueryService {
+
+}
